Make nature configurable only through file name
Using a user map on the psi element, even with stubs,
would only save the nature in the local indices, and
not into version control. If nature should be configurable
ad-hoc, we should persist the set of file which have
a non-default nature. There's no (public) API to do
that (there is something interesting done by the
"force plain text nature" subsystem though), and I
don't care enough to implement that without a demand.
Filename conventions should be observed anyway.